extracting the debug feature state creation to its own factory
